Fix Layer 0 on DesignPerim
Modified SetDecoder to handle decoder 3 as an option.  Layer 0 working now

Need to fix the disabling of certain decoders as it broke when I fixed this layer 0 part

If I use the commented code, which makes sense to me, the entire function breaks out early.  If using other code for decoder 3 only, it turns on but not the same intensity I intended.