Fill Designs working by themselves
The new two fill designs needed to be broken into 4 frames each.  If the
case function only runs one of them, it works fine, but if trying to do
both or the random function to go through all, it hangs and does nothing.

Currently set to run the standard show without fills only.